+++ src/pages/like/components/Like/Like.styles.ts
