+++ src/pages/like/components/Like/Member/LikeMemberCard/index.ts
@@ -1,0 +1,1 @@
+export { LikeMemberCard } from "./LikeMemberCard";
