+++ src/pages/like/components/Like/Owner/LikeOwnerCard/index.ts
@@ -1,0 +1,1 @@
+export { LikeOwnerCard } from "./LikeOwnerCard";
