+++ src/pages/like/components/Like/Owner/index.ts
