+++ src/pages/like/components/Like/index.ts
